Breaking Tritium because breaking Assets AssetsProvider and MapProcessor.
