+++ Sources/Guld/Internal/MapProcessor/MapProcessor.swift
@@ -21,7 +21,7 @@ public extension MapProcessor {
     func process(map: Map) throws -> ProcessedMap {
         print("✨ Processing map \"\(map.basicInformation.name)\"")
         let start = CFAbsoluteTimeGetCurrent()
-        let groundImagesForTiles = try loadGroundImagesForTiles(in: map)
+//        let groundImagesForTiles = try loadGroundImagesForTiles(in: map)
         let riverImagesForTiles = try loadRiverImagesForTiles(in: map)
         let roadImagesForTiles = try loadRoadImagesForTiles(in: map)
         let imageForObjects = try loadObjectImageForObjects(in: map)
@@ -42,10 +42,11 @@ public extension MapProcessor {
         }
         
         func groundImage(at tile: Map.Tile) -> GroundImage {
-            guard let image = groundImagesForTiles[tile] else {
-                fatalError("Expected to always have a ground image.")
-            }
-            return image
+//            guard let image = groundImagesForTiles[tile] else {
+//                fatalError("Expected to always have a ground image.")
+//            }
+//            return image
+            fatalError()
         }
         
         func riverImage(at tile: Map.Tile) -> RiverImage? {
@@ -105,53 +106,56 @@ public extension MapProcessor {
 // MARK: Private
 private extension MapProcessor {
     
-    func _loadImagesForTile<K: DrawableTileLayer>(
-        in map: Map,
-        keyForTile: (Map.Tile) -> K?
-    ) throws -> [Map.Tile: CachedImage<K>] {
-        
-        let tiles = map.world.above.tiles + (map.world.underground?.tiles ?? [])
-        
-        return Dictionary(uniqueKeysWithValues: try tiles.compactMap { tile in
-            guard let key = keyForTile(tile) else {
-                return nil
-            }
-            let image = try assets.loadImage(for: key)
-            return (key: tile, value: image)
-        })
-    }
-    
-    func _loadImagesForTile<K: DrawableTileLayer>(in map: Map, keyBy keyKeyPath: KeyPath<Map.Tile, K>) throws -> [Map.Tile: CachedImage<K>] {
-        try _loadImagesForTile(in: map, keyForTile: { $0[keyPath: keyKeyPath] }  )
-    }
-    
-    func _loadImagesForTile<K: DrawableTileLayer>(in map: Map, maybeKey keyKeyPath: KeyPath<Map.Tile, K?>) throws -> [Map.Tile: CachedImage<K>] {
-        try _loadImagesForTile(in: map, keyForTile: { $0[keyPath: keyKeyPath] }  )
-    }
-    
-    func loadGroundImagesForTiles(in map: Map) throws -> [Map.Tile: GroundImage] {
-        try _loadImagesForTile(in: map, keyBy: \.ground)
-    }
+//    func _loadImagesForTile<K: DrawableTileLayer>(
+//        in map: Map,
+//        keyForTile: (Map.Tile) -> K?
+//    ) throws -> [Map.Tile: CachedImage<K>] {
+//
+//        let tiles = map.world.above.tiles + (map.world.underground?.tiles ?? [])
+//
+//        return Dictionary(uniqueKeysWithValues: try tiles.compactMap { tile in
+//            guard let key = keyForTile(tile) else {
+//                return nil
+//            }
+//            let image = try assets.loadImage(for: key)
+//            return (key: tile, value: image)
+//        })
+//    }
+//
+//    func _loadImagesForTile<K: DrawableTileLayer>(in map: Map, keyBy keyKeyPath: KeyPath<Map.Tile, K>) throws -> [Map.Tile: CachedImage<K>] {
+//        try _loadImagesForTile(in: map, keyForTile: { $0[keyPath: keyKeyPath] }  )
+//    }
+//
+//    func _loadImagesForTile<K: DrawableTileLayer>(in map: Map, maybeKey keyKeyPath: KeyPath<Map.Tile, K?>) throws -> [Map.Tile: CachedImage<K>] {
+//        try _loadImagesForTile(in: map, keyForTile: { $0[keyPath: keyKeyPath] }  )
+//    }
+//
+//    func loadGroundImagesForTiles(in map: Map) throws -> [Map.Tile: GroundImage] {
+//        try _loadImagesForTile(in: map, keyBy: \.ground)
+//    }
     
     func loadRiverImagesForTiles(in map: Map) throws -> [Map.Tile: RiverImage]? {
-        guard case let cache = try _loadImagesForTile(in: map, maybeKey: \.river), !cache.isEmpty else {
-            return nil
-        }
-        return cache
+//        guard case let cache = try _loadImagesForTile(in: map, maybeKey: \.river), !cache.isEmpty else {
+//            return nil
+//        }
+//        return cache
+        fatalError()
     }
     
     func loadRoadImagesForTiles(in map: Map) throws -> [Map.Tile: RoadImage]? {
-        guard case let cache = try _loadImagesForTile(in: map, maybeKey: \.road), !cache.isEmpty else {
-            return nil
-        }
-        return cache
+//        guard case let cache = try _loadImagesForTile(in: map, maybeKey: \.road), !cache.isEmpty else {
+//            return nil
+//        }
+//        return cache
+        fatalError()
     }
     
     func loadObjectImageForObjects(in map: Map) throws -> [Sprite: ObjectImage] {
-        Dictionary<Sprite, ObjectImage>(try map.detailsAboutObjects.objects.compactMap { object in
-            let key = object.attributes.sprite
-            return (key: key, value: try assets.loadImage(sprite: key))
-        }, uniquingKeysWith: { (first, _) in first })
+//        Dictionary<Sprite, ObjectImage>(try map.detailsAboutObjects.objects.compactMap { object in
+//            let key = object.attributes.sprite
+//            return (key: key, value: try assets.loadImage(sprite: key))
+//        }, uniquingKeysWith: { (first, _) in first })
+        fatalError()
     }
 }
