+++ Sources/Guld/Internal/VID/VIDArchiveParser.swift
@@ -9,6 +9,8 @@ import Foundation
 import Util
 import Malm
 
+public typealias ArchiveFile = SimpleFile
+
 internal protocol ArchiveFileCountParser {
     func peekFileEntryCount(of archiveFile: ArchiveFile) throws -> Int
 }
@@ -31,7 +33,6 @@ internal extension VIDArchiveParser {
         archiveFile: ArchiveFile,
         inspector: AssetParsedInspector? = nil
     ) throws -> VIDFile {
-        precondition(archiveFile.kind.isVIDFile)
         
         let reader = DataReader(data: archiveFile.data)
         
@@ -68,7 +69,7 @@ internal extension VIDArchiveParser {
             try reader.seek(to: $0.offset)
             let contents = try reader.read(byteCount: $0.size)
             let fileEntry = VIDFile.FileEntry(
-                parentArchiveName: archiveFile.fileName,
+                parentArchiveName: archiveFile.name,
                 fileName: $0.name,
                 contents: contents
             )
@@ -77,7 +78,7 @@ internal extension VIDArchiveParser {
         }
         
         return .init(
-            archiveKind: archiveFile.kind,
+            archiveName: archiveFile.name,
             entries: entries
         )
     }
