+++ Sources/Guld/Internal/VID/VIDFile.swift
@@ -9,23 +9,18 @@ import Foundation
 import Malm
 
 public final class VIDFile: ArchiveProtocol {
-    public let archiveKind: Archive
+    public let archiveName: String
     public let entries: [FileEntry]
     
     public init(
-        archiveKind: Archive,
+        archiveName: String,
         entries: [FileEntry]
     ) {
-        self.archiveKind = archiveKind
+        self.archiveName = archiveName
         self.entries = entries
     }
 }
 
-public extension VIDFile {
-    var archiveName: String {
-        archiveKind.fileName
-    }
-}
 
 public extension VIDFile {
